requirements.txt
Django==1.9.6
argparse==1.2.1
django-tastypie==0.13.3
python-dateutil==2.5.3
python-mimeparse==1.5.2
six==1.10.0
wsgiref==0.1.2
